Implement New July 2017 LH Export Format, Tighten Regular Expressions
As of July 2017, LH have added a new first column to their export format,
which contains a psedonymised version of the credit card number in the
form "123456xxxxxx1234" (six digits, six lower case "x", 4 digits). This
was added.

Also tightened the regular expressions, which were somewhat lenient. This
was needed since the old REs matching the invoice number would also match
the new credit card number column, as they did not include the string
beginning and end markers to make sure that the field consists of digits
exclusively. This has been addressed.