Add Ability to Process New April 2018 Format
In March/April 2018, DKB updated their systems, resulting in a new
website design, and an all-new CSV transactions export format. This
update adds the ability to process CSV files in the new 2018
format. The workflow remains unchanged, i.e. first process with the
LH_split awk script, then import into MonKey Office.

The new 2018 CSV format no longer conveys the M&M member number, and
hence it is no longer included in the output of the LH_split awk
script. Consequently, the MonKey Office had to be updated also to
check for it, and compare it against the bank account number only if
and when an M&M member number is present in the import file.