Sale Amount Off-By-One For EUR Transactions
On EUR (i.e. non-foreign) transactions, the sale amount was set to the
amount of the previous transaction, instead of the amount of the
current transaction (off-by-one). This is fixed with the present
commit. This bug was uncritical, however, since the sale amount column
is used by the MonKey Office import script for foreign
currency (i.e. non-EUR) transactions only, and the foreign currency
amount is copied to the note field, i.e. did not affect the accounting
results in any way.