[New Feature] Convert Bank X Transactions to DATEV
The new script "bx2datev" converts transactions from the app "Bank
X" (https://www.application-systems.de/bankx/) exported in "komplette
Buchungen" format to DATEV format.

Compared to Bank X's built-in DATEV export the script does not
truncate the booking text, and instead of the other account's number
and BLZ the script inserts the other account's IBAN and BIC.
Unfortunately, Bank X's export format "komplette Buchungen" for the
own account only conveys the name as configured in the GUI by the
user. Consequently, there is no reliable way of inferring the account
number or IBAN from the exported data. Thus, the script needs a little
help for this (see lines 13 through 15 of the script). The recommended
way is to set the environment variable ACCOUNT to the IBAN or (short)
account number.